# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AGR \u2014 Agentic Governance Runtime",
    "description": "Drop-in governance layer for AI agent frameworks. Every agent tool call is evaluated against Cedar policies before execution. Sensitive actions require human approval.\n\n## Authentication\n\nAll endpoints (except `/v1/health`) require `Authorization: Bearer agr_sk_<key>` header.\n\n## Decision Values\n\n- `ALLOW` \u2014 agent may proceed\n- `DENY` \u2014 action is blocked\n- `APPROVAL_REQUIRED` \u2014 action is queued for human approval",
    "contact": {
      "name": "Shreeja AI",
      "url": "https://shreejaai.com/"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/policies": {
      "get": {
        "tags": [
          "policies"
        ],
        "summary": "List Policies",
        "operationId": "list_policies_v1_policies_get",
        "parameters": [
          {
            "name": "active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Active"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(draft|active|archived)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PolicyResponse"
                  },
                  "title": "Response List Policies V1 Policies Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "policies"
        ],
        "summary": "Create Policy",
        "operationId": "create_policy_v1_policies_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/policies/import": {
      "post": {
        "tags": [
          "policies"
        ],
        "summary": "Bulk Import Policies",
        "description": "Bulk import policies from a JSON list.\n\nSet dry_run=true to validate without writing.\nSet overwrite=true to update existing policies with matching names.",
        "operationId": "bulk_import_policies_v1_policies_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/policies/export": {
      "get": {
        "tags": [
          "policies"
        ],
        "summary": "Bulk Export Policies",
        "description": "Export all org policies as a JSON array suitable for re-import.",
        "operationId": "bulk_export_policies_v1_policies_export_get",
        "parameters": [
          {
            "name": "active_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Export only active policies",
              "default": false,
              "title": "Active Only"
            },
            "description": "Export only active policies"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/policies/simulate": {
      "post": {
        "tags": [
          "policies"
        ],
        "summary": "Simulate Policy",
        "description": "Simulate a policy decision without any side effects.\n\nRuns the full Cedar evaluation + risk scoring pipeline against the org's\nactive policies but writes nothing \u2014 no audit event, no approval row,\nno cache entry, no eval_count increment.",
        "operationId": "simulate_policy_v1_policies_simulate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/policies/{policy_id}": {
      "get": {
        "tags": [
          "policies"
        ],
        "summary": "Get Policy",
        "operationId": "get_policy_v1_policies__policy_id__get",
        "parameters": [
          {
            "name": "policy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Policy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "policies"
        ],
        "summary": "Update Policy",
        "operationId": "update_policy_v1_policies__policy_id__patch",
        "parameters": [
          {
            "name": "policy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Policy Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "policies"
        ],
        "summary": "Delete Policy",
        "description": "Soft-delete a policy by archiving it (state=archived, active=False).\n\nM3: audit_events.policy_id holds historical references to policies;\nhard-deleting would leave stale foreign keys in the immutable audit trail.\nArchiving preserves the row (and audit integrity) while stopping the\npolicy from being evaluated in future requests.",
        "operationId": "delete_policy_v1_policies__policy_id__delete",
        "parameters": [
          {
            "name": "policy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Policy Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/policies/{policy_id}/activate": {
      "patch": {
        "tags": [
          "policies"
        ],
        "summary": "Activate Policy",
        "description": "Transition a draft or archived policy to active state.\n\nOnly active policies are evaluated by the Cedar engine.",
        "operationId": "activate_policy_v1_policies__policy_id__activate_patch",
        "parameters": [
          {
            "name": "policy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Policy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/policies/{policy_id}/archive": {
      "patch": {
        "tags": [
          "policies"
        ],
        "summary": "Archive Policy",
        "description": "Archive a policy, removing it from evaluation without hard-deleting it.\n\nArchived policies preserve the audit trail FK integrity and can be\ninspected but not evaluated or updated.",
        "operationId": "archive_policy_v1_policies__policy_id__archive_patch",
        "parameters": [
          {
            "name": "policy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Policy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/policies/{policy_id}/versions": {
      "get": {
        "tags": [
          "policies"
        ],
        "summary": "List Policy Versions",
        "description": "Return the full version history for a policy, newest first.\n\nEach entry is a snapshot taken immediately before a content-changing PATCH.\nThe live policy state is not included \u2014 fetch GET /v1/policies/{id} for that.",
        "operationId": "list_policy_versions_v1_policies__policy_id__versions_get",
        "parameters": [
          {
            "name": "policy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Policy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PolicyVersionResponse"
                  },
                  "title": "Response List Policy Versions V1 Policies  Policy Id  Versions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PolicyResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "org_id": {
            "type": "string",
            "title": "Org Id"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "agent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "level": {
            "type": "string",
            "title": "Level"
          },
          "cedar_rule": {
            "type": "string",
            "title": "Cedar Rule"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "conflicts": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conflicts"
          }
        },
        "type": "object",
        "required": [
          "id",
          "org_id",
          "project_id",
          "agent_id",
          "name",
          "level",
          "cedar_rule",
          "version",
          "active",
          "state",
          "created_at",
          "updated_at"
        ],
        "title": "PolicyResponse"
      },
...
    }
  },
...
}